requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.26.0
pillow>=9.0.0
pandas>=2.0.0
matplotlib>=3.5.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.2.0
tabulate>=0.9.0
h5py>=3.1.0 